big and very epic changes
- add libraries: dfrobot environmental sensor, particulate matter sensor, pid
- improve bluetooth button handling code verbosity
- change gps coord handling to double precision
- add preliminary navigation code
- add some mini mote pin declarations
- add particulate matter sensor code

+++ include/modem.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <Arduino.h>
+
 #define LOG_MODEM 0 // Duplicate all modem communications to Serial for debugging
 
 typedef struct ModemGPSData {
-    float lat, lng;
+    double lat, lng;
     float speed;
     float track;
 } ModemGPSData;
